Fixed indentation issue with .env - Resolves #2

+++ setup.sh
@@ -14,45 +14,67 @@ printMessage() {
 
 printMessage "########### Welcome to BoltCTF Setup #############"
 
-printMessage "\nChecking for required programs and utilities..."
+printMessage "\nCHECKING: required programs and utilities..."
 checkProgram php
 checkProgram composer
 checkProgram npm
 
-printMessage "\nFetching and installing all dependencies..."
+printMessage "\nINSTALLING: dependencies..."
 composer install --optimize-autoloader
 npm install
 
-printMessage "\nProcessing .env file..."
-read -p "Would you like to keep your old .env file? (y/n) [y]" -n 1 -r
-if [[ $REPLY =~ ^[Nn]$ ]]
-then
-    rm .env
-    printMessage "\nProvide configuration for the database..."
-    read -p "Enter Your Database Name: "  dbname
-    read -p "Enter Your Database User Name: "  dbuser
-    read -p "Enter Your Database Password: "  dbpassword
-    read -p "Enter Your Database Host name: "  dbhostname
+printMessage "\nCHECKING: .env file..."
+if [ -e .env ]; then
+  read -p "Would you like to keep your old .env file? (y/n) [y]" -n 1 -r
+  if [[ $REPLY =~ ^[Nn]$ ]]
+  then
+      rm .env
+      printMessage "Provide configuration for the database..."
+      read -p "Enter Your Database Name: "  dbname
+      read -p "Enter Your Database User Name: "  dbuser
+      read -p "Enter Your Database Password: "  dbpassword
+      read -p "Enter Your Database Host name: "  dbhostname
 
-    echo -e "APP_ENV=production\nAPP_DEBUG=false\nAPP_KEY=\n" >> .env
+######### Excuse the weird indentation (silly cat!) #########
+cat > .env << EOF
+APP_ENV=production
+APP_TZ=America/Chicago
+APP_DEBUG=false
+APP_KEY=
 
-    echo -e "DB_HOST=$dbhostname\nDB_DATABASE=$dbname\nDB_USERNAME=$dbuser\nDB_PASSWORD=$dbpassword\n" >> .env
+DB_HOST=$dbhostname
+DB_DATABASE=$dbname
+DB_USERNAME=$dbuser
+DB_PASSWORD=$dbpassword
 
-    echo -e "CACHE_DRIVER=file\nSESSION_DRIVER=file\nQUEUE_DRIVER=sync\n" >> .env
+CACHE_DRIVER=file
+SESSION_DRIVER=file
+QUEUE_DRIVER=sync
 
-    echo -e "REDIS_HOST=localhost\nREDIS_PASSWORD=null\nREDIS_PORT=6379\n" >> .env
+REDIS_HOST=localhost
+REDIS_PASSWORD=null
+REDIS_PORT=6379
 
-    echo -e "MAIL_DRIVER=smtp\nMAIL_HOST=mailtrap.io\nMAIL_PORT=2525\nMAIL_USERNAME=null\nMAIL_PASSWORD=null\nMAIL_ENCRYPTION=null\n" >> .env
+MAIL_DRIVER=smtp
+MAIL_HOST=mailtrap.io
+MAIL_PORT=2525
+MAIL_USERNAME=null
+MAIL_PASSWORD=null
+MAIL_ENCRYPTION=null
+MAIL_FROM_ADDRESS=no-reply@example.com
+MAIL_FROM_NAME=no-reply
+
+EOF
+######### Excuse the weird indentation (silly cat!) #########
+
+  fi
 fi
 
-printMessage "\nGenerating encryption keys.."
-printMessage "Type yes if running for the first time, else press enter key to skip"
+printMessage "\nENCRYPTION KEYS: Type yes if running for the first time, else press enter key to skip"
 php artisan passport:keys
 php artisan key:generate
-printMessage "\nYou can edit your .env file if anything changes in future."
 
-printMessage "\nBuilding asset libraries"
-npm rebuild node-sass --force
+printMessage "\nBUILDING: asset libraries"
 npm run prod
 
 printMessage "\nRunning optimizer and migrator."
@@ -60,7 +82,7 @@ printMessage "\nRunning optimizer and migrator."
 php artisan optimize
 php artisan cache:clear
 
-printMessage "\nDATABASE MIGRATION : Type yes if running for the first time else press enter key to skip"
-php artisan migrate:refresh --seed
+printMessage "\nType yes if running for the first time else press enter key to skip"
+php artisan migrate --seed
 
-printMessage "\nAll Done. Run 'php artisan serve' to launch the app or check the README file for NGINX server setup."
+printMessage "\nAll Done. Run 'php artisan serve' to launch the app."
